week02/day13/requirements.txt
faiss-cpu==1.8.0
numpy>=1.23
gradio>=4.0
